bpf: initial prefilter work for ddos mitigation
This work enables Cilium to make use of XDP for prefiltering
and thus allows early dropping traffic from blacklisted CIDRs.

The agent gets two command line options:

  --prefilter-device <netdev> --prefilter-mode {native|generic}

First argument is the install device for the XDP program, and
the second argument defines whether native XDP or generic XDP
should be used. If nothing specified, we default to native XDP.

The following commands can be used from cilium cli, e.g.:

  cilium prefilter list
  -> Lists the current revision number and all installed CIDRs.

  cilium prefilter update --revision=X --cidr=::1/128,::2/128
  -> Installs the provided list of CIDRs for the specified
     revision number. The revision is bumped after the install
     was successful.

  cilium prefilter delete --revision=Y --cidr=::1/128,::2/128
  -> Deletes the provided list of CIDRs for the specified
     revision number. The revision is bumped after the delete
     was successful.

Apart from the cli, the http/rest API can also be used directly
for managing CIDRs for XDP. LPM map is also possible, but
currently not activated due to missing get_next_key() from
kernel side. Idea is that once we see one of the CIDRs to be
installed having a prefix other than /32 or /128, then we add
a LPM on the fly and install the CIDR there.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>